# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Taiko Prover Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://community.taiko.xyz/",
            "email": "info@taiko.xyz"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/taikoxyz/taiko-client/blob/main/LICENSE.md"
        },
        "version": "1.0"
    },
    "paths": {
        "/assignment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Try to accept a block proof assignment",
                "operationId": "create-assignment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/prover_server.ProposeBlockResponse"
                        }
                    },
                    "422": {
                        "description": "prover does not have capacity",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get current prover server status",
                "operationId": "get-status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/prover_server.Status"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "prover_server.ProposeBlockResponse": {
            "type": "object",
            "properties": {
                "prover": {
                    "type": "string"
                },
                "signedPayload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "prover_server.Status": {
            "type": "object",
            "properties": {
                "currentCapacity": {
                    "type": "integer"
                },
                "maxExpiry": {
                    "type": "integer"
                },
                "minOptimisticTierFee": {
                    "type": "integer"
                },
                "minPseZkevmTierFee": {
                    "type": "integer"
                },
                "minSgxTierFee": {
                    "type": "integer"
                }
            }
        }
    }
}